+++ whiteboard.h
@@ -66,7 +66,7 @@ janus_whiteboard *janus_whiteboard_create(const char *dir, const char *filename,
  * @param[in] buffer The frame data to save
  * @param[in] length The frame data length
  * @returns 0 in case of success, a negative integer otherwise */
-int janus_whiteboard_save_package(janus_whiteboard *whiteboard, char *buffer, uint length, uint8_t **out);
+int janus_whiteboard_save_package(janus_whiteboard *whiteboard, char *buffer, uint16_t length, uint8_t **out);
 
 uint8_t *janus_whiteboard_current_scene_data(janus_whiteboard *whiteboard, int *size);
 /*! \brief Close the whiteboard
